+++ pages/_app.js
@@ -1,7 +1,7 @@
 import GlobalStyle from "../styles";
 import Head from "next/head";
 import useSWR from "swr";
-import PosesList from "../components/PosesList";
+import RandomPoses from "./random-poses";
 
 const fetcher = (url) => fetch(url).then((res) => res.json());
 
@@ -22,6 +22,7 @@ export default function App({ Component, pageProps }) {
         <title>Capstone Project</title>
       </Head>
       <Component {...pageProps} poses={poses} />
+      <RandomPoses poses={poses} />
     </>
   );
 }
